put formatting stuff in its own module
