+++ crates/re_arrow_store/src/lib.rs
@@ -16,6 +16,7 @@
 
 mod arrow_util;
 mod store;
+mod store_format;
 mod store_gc;
 mod store_read;
 mod store_sanity;
